bot/pyproject.toml
[project]
name = "vela"
version = "0.1.0"
description = "A Discord bot built with discord.py"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.3",
    "ollama>=0.6.0",
    "pydantic>=2.11.10",
    "pylance>=0.38.0",
    "requests>=2.32.5",
    "ruff>=0.13.3",
    "toml>=0.10.2",
]
